10 sum all the light curves (#13)
* Create Light curve script, implement sum of LC of different gbm detectors.

* Do not ignore .dat light curve files.

* make some curves

* Minor changes. Light curves up to 16.

* Unmask relevant catalog columns

* Remove .unmasked from write_light_curve script.

* Untrack files in .gitignore

* Merging strategy failed for several bugs of the GBM Data Tools.
If we Energy slice a TTE (counts in first, last channels=0) the background fitter is going to fail for a np.linalg Error: covariance matrices are going to be singular.
We must then perform the energy slicing after the Background Fitting, when turning CSPECS into TimeBins.
This can be done if we merge the NaIs only, that have similar energy ranges and channels.
It cannot be done if we want to merge the NaIs and the BGO, since the counts from all are merged in the same 128 channels with no respect for the energy range of the original detector.
So we can't slice before the merging because the Background Fitter will fail, and we can't slice after the merging because the original energies are lost.
Possible solution: merge only the NaIs. The NaI+BGO can be done with TimeBins.sum, altough this means that if the same events are seen by the BGO and the NaIs we are going to count them twice.

* Implement solutions from last commit: NaIs are merged, then sliced after the Background fit.
Then after making the TimeBins we sum the BGO with the merged NaIs.
Graphical output improved with peak selection.